Added collector parameters and modules
