+++ src/components/board.js
@@ -1,0 +1,8 @@
+const templateBoard = () => {
+  return `<section class="board container">
+   <div class="board__tasks">
+   </div>
+   </section>`;
+};
+
+export {templateBoard as board};
