+++ src/components/loadMore.js
@@ -1,0 +1,5 @@
+const templateLoadMore = () => {
+  return `<button class="load-more" type="button">load more</button>`;
+};
+
+export {templateLoadMore as loadMore};
